Erase state cookie at the path it was set.
Otherwise it won’t get deleted except if we’re at / which is unlikely.
This was manifesting as a bug where if you’d saved an object / created
a new object of a particular model once via the admin in a session
(setting the state cookie to 1) any subsequent autosave data for that
object / new object of that model was being deleted rather than offered
for recovery (because on detecting the state cookie as 1 the first time,
and clearing the local storage with the autosave data, the state cookie
wasn’t being erased — meaning on every subsequent page load the same thing would happen).

Added a comment of the strategy being used for the state cookie to make
it easier to maintain in future.